==> notebooks/stage_1.ipynb <==
# Eksploracja Danych - Projekt
Tomasz Kiljańczyk (136257)

Wojciech Lulek (136280)

# Załadowanie bibliotek i danych

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os

data_dir = './data'

us_trending_df = pd.read_csv(os.path.join(data_dir, 'USData80.csv'), encoding='cp1250', lineterminator='\n')
us_trending_df = us_trending_df.iloc[:, 1:]
us_trending_df = us_trending_df.rename(columns={'description\r': 'description'})
us_trending_df['description'] = us_trending_df['description'].str.rstrip('\r')
us_trending_df

Dislikes nie jest już możliwe do pobrania, więc można je od razu usunąć

us_trending_df.drop(columns=['dislikes'], axis=1, inplace=True)

# Analiza danych

## Analiza wstępna

us_trending_df.shape

us_trending_df.info()

CategoryId zawiera NaNy, trzeba się ich pozbyć

us_trending_df['categoryId'].replace(np.NaN, 0.0, inplace=True)

CategoryId jest typu float64. Żeby zamienić je na stringa trzeba je zmienić wpierw na int64

us_trending_df['categoryId'] = us_trending_df['categoryId'].astype(np.int64)

Kolumny opisujące daty są typu object.
Można je zrzutować na typ datetime64.
categoryId należy zrzutować na typ string, ponieważ nie chcemy go traktować jako dane numeryczne.

us_trending_df = us_trending_df.astype({
    'trending_date': np.datetime64,
    'publishedAt': np.datetime64,
    'categoryId': str
})
us_trending_df.info()

W kolumnie description występują brakujące wartości.
Można by je uzupełnić pustym stringiem albo spacją.

us_trending_df['description'].replace(np.NaN, '', inplace=True)
us_trending_df.info()

Przy ręcznej analizie znaleziono też tagi `[None]`.
Chyba najlepiej zamienić je na pusty string.

us_trending_df['tags'].replace('[None]', '', inplace=True)

us_trending_df.nunique()

W policzonych unikatowych wartościach widać, że unikatowych wartości video_id jest mniej niż ilość wszystkich rzędów w pliku.

Wynika to z faktu, że w danych znajdują się zrzuty listy trending z wielu dni.
Jeden filmik mógł być trending przez parę dni.

Występują rzędy z view_count = 0. Najlepiej się ich pozbyć

us_trending_df.drop(us_trending_df[us_trending_df['view_count'] == 0].index, inplace=True)

## Wizualizacje

corrMatrix = us_trending_df.corr(method='spearman')
corrMatrix = corrMatrix.round(4)

mask = np.triu(np.ones_like(corrMatrix, dtype=bool))
np.fill_diagonal(mask, False)

fig, ax = plt.subplots(figsize=(10, 10))
sns.heatmap(corrMatrix, mask=mask, annot=True, fmt='g', ax=ax)
plt.show()

Używamy metryki Spearmana, ponieważ to umożliwia odkrycie nieliniowych zależności między cechami.

Zależności większe niż |0.5|:
* view_count:
    * comment_count
    * dislikes
    * likes
* likes:
    * comment_count
    * dislikes
* dislikes:
    * comment_count

Z wypisanych zależności wynika, że im więcej osób ogląda filmik, tym więcej jest komentarzy, polubień i antypolubień.
To jest logiczna konsekwencja wpływu oglądalności na te cechy.

Polubienia można zastąpić cechą stosunek polubień do liczby wyświetleń.

us_trending_df['likes_to_view_count_ratio'] = us_trending_df['likes'] / us_trending_df['view_count']
us_trending_df['likes_to_view_count_ratio']

us_trending_df['likes_to_view_count_ratio'].hist(log=True)

Liczbę komentarzy można zamienić na stosunek komentarzy do liczby obejrzeń.
Należy pamiętać, że stosunek ten może być większy niż 1 (bo jeden użytkownik może komentować wiele razy).

us_trending_df['comment_count_to_view_count_ratio'] = us_trending_df['comment_count'] / us_trending_df['view_count']
us_trending_df['comment_count_to_view_count_ratio']

us_trending_df['comment_count_to_view_count_ratio'].hist(log=True)

us_trending_df['view_count'].hist(log=True)

us_trending_df['comment_count'].hist(log=True)

us_trending_df['likes'].hist(log=True)

## Analiza kategorii

import json

with open(os.path.join(data_dir, 'US_category_id.json'), mode='r') as file:
    us_category_id_dict = json.loads(file.read())

us_categories = {}
for category in us_category_id_dict['items']:
    us_categories[category['id']] = category['snippet']['title']

us_categories

len(us_categories)

Z analizy wynika, że używane jest 15 kategorii, a w pliku JSON dostarczone zostało 32 kategorie.
17 kategorii jest nieużywanych.

us_trending_df['category'] = us_trending_df['categoryId']
us_trending_df['category'] = us_trending_df['category'].apply(us_categories.get).replace(np.NaN, '')
us_trending_df['category']

category_counts = us_trending_df.groupby(["category"]).size().reset_index(name='counts')
category_counts

category_counts.plot(x='category', y='counts', kind='bar', logy=True)

## Analiza tytułów

from sklearn.feature_extraction.text import CountVectorizer

vectorizer = CountVectorizer(strip_accents='unicode', stop_words='english')

titles_with_descriptions = us_trending_df['title'].str.cat(us_trending_df['description'], sep=' ')
titles_with_descriptions

titles_vectorized_csr = vectorizer.fit_transform(titles_with_descriptions)
token_count_sum = titles_vectorized_csr.sum(axis=0).A1
tokens = vectorizer.get_feature_names_out()
df_dict = {'token': tokens, 'count': token_count_sum}
token_count_df = pd.DataFrame(df_dict)
token_count_df

top_words = token_count_df.sort_values(by=['count'], ascending=False).head(150).reset_index(drop=True)
print(top_words.to_string())

for word in top_words['token']:
    us_trending_df[f'word_{word}'] = titles_with_descriptions.str.contains(word, regex=False)

us_trending_df.head()

Notatka:
* Można wpisać słowa jako cechy i potem usuwać metodami z wykładu
* Przydałby się ranking informatywności.
* Można użyć słownika z kategoriami

## Analiza tagów

vectorizer = CountVectorizer(strip_accents='unicode', token_pattern=r'[^|]+')

tags_vectorized_csr = vectorizer.fit_transform(us_trending_df['tags'])
token_count_sum = tags_vectorized_csr.sum(axis=0).A1
tokens = vectorizer.get_feature_names_out()

df_dict = {'token': tokens, 'count': token_count_sum}
token_count_df = pd.DataFrame(df_dict)
token_count_df

top_tags = token_count_df.sort_values(by=['count'], ascending=False).head(100).reset_index(drop=True)
print(top_tags.to_string())

for tag in top_tags['token']:
    us_trending_df[f'tag_{tag}'] = us_trending_df['tags'].str.contains(tag, regex=False)

us_trending_df.head()

# Inne atrybuty

## Nowe atrybuty na podstawie atrybutów tekstowych

us_trending_df['title_length'] = us_trending_df['title'].str.len()
us_trending_df['title_length'].hist()

us_trending_df['title_capital_letters'] = us_trending_df['title'].str.count(r'[A-Z]') / us_trending_df['title_length']
us_trending_df['title_capital_letters'].hist()

PUNCTUATION_REGEX = r'[!"#$%&\'()*+,-./:;<=>?@[\]^_`{|}~]'

us_trending_df['title_punctuation'] = us_trending_df['title'].str.count(PUNCTUATION_REGEX) / us_trending_df['title_length']
us_trending_df['title_punctuation'].hist()

us_trending_df['description_height'] = us_trending_df['description'].str.count(r'[\r\n]')
us_trending_df['description_height'].hist(log=True)

us_trending_df['description_length'] = us_trending_df['description'].str.len()
us_trending_df['description_length'].hist()

us_trending_df['description_link_count'] = us_trending_df['description'].str.count(r'://')
us_trending_df['description_link_count'].hist(log=True)

Notatka: można podzielić linki na kategorie

us_trending_df['tag_count'] = us_trending_df['tags'].apply(lambda tags: 0 if tags == '' else tags.count('|') + 1)
us_trending_df['tag_count'].hist()

## Nowe atrybuty na podstawie dat

us_trending_df['publishedAtHour'] = us_trending_df['publishedAt'].dt.hour
us_trending_df['publishedAtHour'].hist()

us_trending_df['publishedAtDay'] = us_trending_df['publishedAt'].dt.weekday
us_trending_df['publishedAtDay'].hist()

us_trending_df['publishedAtMonth'] = us_trending_df['publishedAt'].dt.month
us_trending_df['publishedAtMonth'].hist()

Notatka:
* Warto się przyjrzeć miesiącom publikacji

from pandas.tseries.holiday import USFederalHolidayCalendar

cal = USFederalHolidayCalendar()
min_date = us_trending_df['publishedAt'].min()
max_date = us_trending_df['publishedAt'].max()

holidays = cal.holidays(start=min_date, end=max_date)
us_trending_df['publishedAtHoliday'] = us_trending_df['publishedAt'].apply(lambda date: date in holidays)
us_trending_df[us_trending_df['publishedAtHoliday']]

Filmiki z listy trending nie były publikowane z święta.

Notatka:
* można spróbować dzień przed/po świętach

Notatka końcowa:
* Warto sprawdzić rozkłady dla różnych kategorii filmików

# Usuwanie kolumn

us_trending_df.info()

* Nie da się wykorzystać
  * Nieprzetworzonych tagów
  * Nieprzetworzonych opisów
  * Nieprzetworzonych tytułów
  * video_id
  * channelId - niekoniecznie
  * channelTitle - niekoniecznie
  * trending_date, bo niewystępuje w danych nie-trending
  * nieprzetworzone publishedAt

* Można zamienić na coś innego:
  * likes i dislikes można zamienić na ilość reakcji i procent likes

Notatka:
* channelId/Title można użyć do sprawdzania kategorii
* Trending date może mieć znaczenie na etapie uczenia (np. w jakim miesiącu są trending).

us_trending_clean_df = us_trending_df.copy()
us_trending_clean_df.drop(["title", "tags", "description"], axis=1, inplace=True)
us_trending_clean_df.info()

us_trending_clean_df

us_trending_clean_df.to_csv("./data/stage_1_us_trending.csv", index=False)
us_trending_clean_df.shape
