==> notebooks/stage_3_download_ground_truth.ipynb <==
# Eksploracja Danych - Projekt
Tomasz Kiljańczyk (136257)

Wojciech Lulek (136280)

import pandas as pd

us_trending_df = pd.read_csv("./data/stage_1_us_trending.csv")
us_trending_df

video_ids = us_trending_df[us_trending_df['categoryId'] == 0]['video_id'].unique()
video_ids

print(len(video_ids))

from notebooks.youtube_api import download_video_categories

category_ids = download_video_categories(video_ids)

len(category_ids)

category_ids_list = []
selected_video_ids = us_trending_df['video_id'].isin(set(list(video_ids)))

for video_id in us_trending_df.loc[selected_video_ids, 'video_id']:
    category_ids_list.append(category_ids.get(video_id, 0))

us_trending_df.loc[selected_video_ids, 'categoryId'] = category_ids_list

us_trending_df[us_trending_df['categoryId'] == 0]

us_trending_df.to_csv("data/stage_3_ground_truth.csv", index=False)
