==> notebooks/stage_3_feature_importance.ipynb <==
# Eksploracja Danych - Projekt
Tomasz Kiljańczyk (136257)

Wojciech Lulek (136280)

Read DF's

import numpy as np
import pandas as pd

us_trending_df = pd.read_csv("./data/stage_1_us_trending.csv")
color_text_df = pd.read_csv("./data/stage_2_thumbnail_colors_and_text.csv")
face_df = pd.read_csv("./data/stage_2_thumbnail_face.csv")
objects_df = pd.read_csv("./data/stage_2_thumbnail_objects_vectors.csv")

print(us_trending_df.shape, color_text_df.shape, face_df.shape, objects_df.shape)

Merge DF's

df_all = pd.concat([us_trending_df, color_text_df, face_df, objects_df], join='inner', axis=1)
df_all.drop('category', axis=1, inplace=True)
df_all.head()

from sklearn.preprocessing import LabelEncoder, Normalizer

encoders = dict()
num_cols = []
for name, type in zip(df_all.columns, df_all.dtypes):
    if type == object:
        encoders[name] = LabelEncoder()
        df_all[name] = encoders[name].fit_transform(df_all[name])

    if type in [np.int64, np.int32, np.float64]:
        num_cols.append(name)

num_cols.remove('categoryId')

df_all[num_cols] = Normalizer().fit_transform(df_all[num_cols])
df_all.dtypes

from featurewiz import featurewiz

df_unlbl = df_all[df_all['categoryId'] == 0]
df_lbl = df_all[df_all['categoryId'] != 0]

f_names, a = featurewiz(df_lbl, target='categoryId', corr_limit=0.70, verbose=2)

f_names

df_all[f_names + ['categoryId']].to_csv('./data/stage_3_feature_importance.csv', index=False)
