==> notebooks/stage_5.ipynb <==
# Eksploracja Danych - Projekt
Tomasz Kiljańczyk (136257)

Wojciech Lulek (136280)

import pandas as pd

us_trending_df = pd.read_csv("./data/stage_1_us_trending.csv")
us_trending_df

us_trending_df['publishedAt'].min()

us_trending_df['publishedAt'].max()

us_trending_df['view_count'].hist(log=True)

view_count_threshold = us_trending_df.view_count.quantile(q=0.1)
view_count_threshold
